Replace CLINT with preliminary ACLINT support
Based on the implementation of CLINT, a preliminary ACLINT is
implemented, including the basic logic for operating `mtimer`, `mswi`,
and `sswi`. CLINT was replaced with ACLINT, and the old CLINT
implementation was removed entirely.

Currently, due to the lack of implementation, the introduced ACLINT uses
only supervisor-level IPI. Therefore, although the logic for mswi is
implemented, it is not being used at the moment.

The implementation can be tested by `make check SMP=n`, where n is the
number of harts you want to simulate.